configs: add RaspberryPi defconfig
This is a default configuration for the RaspberryPi, tested and functionnal.

The kernel used is the one developped by the RaspberryPi foundation as it's not
fully supported currently in mainline kernel.
The configuration used for the kernel is the default bcmrpi.

For more info, please find board/raspberrypi/README.txt

Signed-off-by: Maxime Hadjinlian <maxime.hadjinlian@gmail.com>
Acked-by: Samuel Martin <s.martin49@gmail.com>